COMP: Place python shared libs as subpackage of itk
Variables that are automatically set should be present in the
CACHE for easier reviewing.

Add sanity check for required file needed for testing When testing, the WrapITK.pth file MUST exist.
