+++ Wrapping/Generators/Python/itk/support/ModuleConfig.py.in
