+++ Wrapping/Generators/Python/itk/support/itkBase.py
@@ -23,8 +23,8 @@ from sys import stderr as system_error_stream
 # Required to work around weird import error with xarray
 import pkg_resources
 
-import itkConfig
-import itkTemplate
+from itk.conf import itkConfig
+from itk.support.itkTemplate import itkTemplate
 
 
 def create_itk_module(name):
