+++ Wrapping/Generators/Python/itk/support/itkBuildOptions.py.in
@@ -1,5 +1,5 @@
-import itkTemplate
-import itkTypes
+from itk.support import itkTypes
+from itk.support.itkTemplate import itkTemplate
 
 DIMS = [int(s) for s in "@ITK_WRAP_IMAGE_DIMS@".split(";") if s]
 USIGN_INTS = [getattr(itkTypes, s) for s in "@WRAP_ITK_USIGN_INT@".split(";") if s]
