+++ Wrapping/Generators/Python/itk/support/itkTypes.py
